fix(studio): initialize Sentry on the TanStack build (captures were silent no-ops)
Nothing loads Next's convention files (instrumentation-client.ts) under
TanStack Start, so the TanStack runtime never called Sentry.init — every
Sentry.captureException there (including the globalErrorBoundary and
routerErrorComponent captures in routes/__root.tsx) was a silent no-op.

- Extract the entire client config (DSN/environment envs, sample rates,
  integrations, allowUrls, beforeBreadcrumb, beforeSend filters,
  ignoreErrors) from instrumentation-client.ts into a shared builder,
  lib/sentry-client-options.ts, so the two runtimes cannot drift.
  instrumentation-client.ts now just calls Sentry.init with it.
- Add sentry.tanstack.ts, called from getRouter() (router.tsx) — the
  earliest point with access to the router instance — wiring
  tanstackRouterBrowserTracingIntegration for pageload/navigation spans.
  Guarded to the browser (getRouter also runs during SSR/prerender) and
  idempotent. Not named *.client.* on purpose: TanStack Start's
  import-protection denies such modules in the server bundle.
- thirdPartyErrorFilterIntegration stays Next-only: the Vite build runs no
  Sentry bundler plugin, so frames carry no applicationKey metadata and the
  integration would tag EVERY event third_party_code=true — beforeSend
  would then drop all non-error-boundary events. Follow-up: annotate via
  @sentry/vite-plugin moduleMetadata, then enable it.
- vite.config.ts: inline `undefined` for unset NEXT_PUBLIC_SENTRY_DSN /
  NEXT_PUBLIC_SENTRY_ENVIRONMENT so self-hosted builds without a DSN don't
  ship a literal `process.env.*` (undeclared `process` throws in the
  browser). With no DSN the init is a clean no-op.
- Move instrumentation-client.test.ts to lib/sentry-client-options.test.ts
  and add coverage for the integration gating and Next/TanStack parity.
