+++ apps/studio/.github/eslint-rule-baselines.json
@@ -556,7 +556,6 @@
       "hooks/misc/withAuth.tsx": 1,
       "hooks/ui/useClickedOutside.ts": 2,
       "hooks/ui/useFlag.ts": 1,
-      "instrumentation-client.ts": 3,
       "lib/ai/generate-assistant-response.ts": 2,
       "lib/ai/model.ts": 2,
       "lib/ai/model.utils.ts": 3,
@@ -571,6 +570,7 @@
       "lib/pg-format.ts": 6,
       "lib/profile.tsx": 1,
       "lib/role-impersonation.ts": 1,
+      "lib/sentry-client-options.ts": 3,
       "lib/telemetry/track.ts": 1,
       "pages/_app.tsx": 1,
       "pages/api/ai/feedback/rate.ts": 2,
